# energy_communities_api/doc/openapi.yml
openapi: 3.0.3
info:
  title: Energy selfconsumption API-REST
  description: API-REST documentation for selfconsumption module.
  termsOfService: https://somcomunitats.coop/avis-legal/
  contact:
    email: a@a.foo
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  version: 14.0.1.0.0
servers:
  - url: https://apifoo.org
tags:
  - name: selfconsumption_projects
    description: All information about selfconsumption projects in an energy community
paths:
  /energy_selfconsumption/projects:
    get:
      tags:
        - selfconsumption_projects
      summary: Obtain all selfconsumption in a energy community
      description:
        Get a list of selfconsumption projects that belongs to the energy community of
        the user
      responses:
        200:
          $ref: "#/components/responses/SelfConsumptionProjectsInfoResponse"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
      security:
        - odooCEAuth: []
  /energy_selfconsumption/projects/{project_code}:
    get:
      tags:
        - selfconsumption_projects
      summary: Find selfconsumption project by cau
      description: Returns the project that match with cau
      parameters:
        - $ref: "#/components/parameters/ProjectcodeParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        200:
          $ref: "#/components/responses/SelfConsumptionSingleProjectInfoResponse"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
      security:
        - odooCEAuth: []
  /energy_selfconsumption/projects/{project_code}/members:
    get:
      tags:
        - selfconsumption_projects
      summary:
        Find all members of an energy community that belongs to an especific
        selfconsumption project
      description:
        Get a list of members of an  energy community that aire involve in a
        selfconsumption project - installation defined by cau
      parameters:
        - $ref: "#/components/parameters/ProjectcodeParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        200:
          $ref: "#/components/responses/SelfConsumptionSingleProjectMembersResponse"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
      security:
        - odooCEAuth: []
components:
  securitySchemes:
    odooCEAuth:
      type: http
      scheme: bearer
      bearerFormat: api-key
  parameters:
    ProjectcodeParam:
      name: project_code
      in: path
      description: project code of an specific installation
      required: true
      schema:
        type: string
      example: ES1234123456789012JY1FA000
    PageParam:
      name: page
      in: query
      description: >
        Actual page when results are paginted
      required: false
      style: form
      schema:
        type: integer
        example: 2
    LimitParam:
      name: limit
      in: query
      description: >
        Number of results per page when a search es paginated
      required: false
      style: form
      schema:
        type: integer
        example: 50
  responses:
    BadRequestResponse:
      description: This response is generated if you have built incorrectly your request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badrequest-response-example:
              $ref: "#/components/examples/badrequest-response-example"
    UnauthorizedResponse:
      description:
        If you try to request some endpoint, and you are not authorized, the api will
        answer you with this response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized-response-example:
              $ref: "#/components/examples/unauthorized-response-example"
    SelfConsumptionProjectsInfoResponse:
      description:
        Returns a set of selfconsumption projects with the number of results and the
        corresponding data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProjectsInfoListResponse"
          examples:
            contracts-response-example:
              $ref: "#/components/examples/selfconsumption_projectsinfo-response-example"
    SelfConsumptionSingleProjectInfoResponse:
      description: Returns a single selfconsumption project information.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SingleProjectInfoResponse"
          examples:
            single-contract-response-example:
              $ref: "#/components/examples/selfconsumption_singleprojectinfo-response-example"
    SelfConsumptionSingleProjectMembersResponse:
      description: Returns the member list of a especific selfconsumption installation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SingleProjectMembersResponse"
          examples:
            single-contract-response-example:
              $ref: "#/components/examples/selfconsumption_singleprojectmembers-response-example"
  schemas:
    SelfConsumptionProjectInfo:
      type: object
      properties:
        project_code:
          type: string
          description: installation cau
          example: ""
        project_name:
          type: string
          description: Name of the project
          example: Super chachi solar and wind installation
        energy_community_id:
          type: string
          description: ID of energy community that belongs this installation
          example: ""
        energy_community_name:
          type: string
          description: Name of the energy community that belongs this installation
          example: "GaresBide"
        power:
          type: integer
          description: Total power of the installation, in kW
          example: 6.2
    SelfConsumptionSingleProjectMembers:
      type: object
      properties:
        supply_point:
          type: string
          description: supply_point code for this member
        supply_point_address:
          type: string
          description: Addess of supply_point
        supply_point_postalcode:
          type: string
          description: Postalcode of supply_point
        supply_point_city:
          type: string
          description: city / town or village of this supply_point
        supply_point_state:
          type: string
          description: State of this supply_point
        beta:
          type: integer
          description: Sharing coeficient for this supply_point
        owner_name:
          type: string
          description: Name of the owner of the supply_point
        owner_surnames:
          type: string
          description: Surname of the owner
        owner_vat:
          type: string
          description: ID of the owner
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        description:
          type: string
          description: Error description
    # ============================
    # Request and Response schemas
    # ============================
    BaseListResponse:
      type: object
      properties:
        total_results:
          type: integer
          description: Total results in general for this request
        count:
          type: integer
          description: Number of results on this response
        page:
          type: integer
          description: Number of the page in where whe are
    PaginationLinks:
      type: object
      properties:
        self:
          type: string
          description: Actual results url
        next_page:
          type: string
          description: Url of next results. If null you are in the las page
        previous_page:
          type: string
          description: Url of the previous results. If null you are in the first page
    ErrorResponse:
      allOf:
        - type: object
          properties:
            error:
              $ref: "#/components/schemas/Error"
    ProjectsInfoListResponse:
      allOf:
        - $ref: "#/components/schemas/BaseListResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SelfConsumptionProjectInfo"
            _links:
              $ref: "#/components/schemas/PaginationLinks"
    SingleProjectInfoResponse:
      allOf:
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SelfConsumptionProjectInfo"
            _links:
              $ref: "#/components/schemas/PaginationLinks"
    SingleProjectMembersResponse:
      allOf:
        - $ref: "#/components/schemas/BaseListResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SelfConsumptionSingleProjectMembers"
            _links:
              $ref: "#/components/schemas/PaginationLinks"
  examples:
    selfconsumption_projectsinfo-response-example:
      value:
        total_results: 10
        count: 2
        page: 2
        data:
          - project_code: "ES1234123456789012JY1FA000"
            project_name: "Taulada Escola Oficis"
            energy_community_id: 1
            energy_community_name: "Comunitat energètica Sol-ete"
            power: 5.6
          - project_code: "ES0288529987082650SF1FA000"
            project_name: "Autoconsum comunitat veins carrer industria"
            energy_community_id: 1
            energy_community_name: "Comunitat energètica Sol-ete"
            power: 7.6
        _links:
          self: https://somcomunitats.coop/api/energy_selfconsumption/projects?page=2&size=2
          previous_page: https://somcomunitats.coop/api/energy_selfconsumption/projects?page=2&size=0
          next_page: https://somcomunitats.coop/api/energy_selfconsumption/projects?page=3&size=2
    selfconsumption_singleprojectinfo-response-example:
      value:
        data:
          cau: "ES0021158966826855BSA008"
          project_name: "Taulada den Xurri"
          ce_id: "2"
          ce_name: "Energia caSOLana"
          power: 5.6
        _links:
          self: https://somcomunitats.coop/api/energy_selfconsumption/projects/ES0021158966826855BSA008
          previous_page: null
          next_page: null
    selfconsumption_singleprojectmembers-response-example:
      value:
        total_results: 10
        count: 2
        page: 1
        data:
          - supply_point: "ES0021158966826855BS"
            supply_point_address: Travesía iglesia, 50
            supply_point_postalcode: 08742
            supply_point_town: Sant Quintí de Mediona
            supply_point_state: Barcelona
            distribution_coefficient: 0.5
            owner_name: Maria Amparo
            owner_surname1: Caballé
            owner_surname2: Roig
            owner_vat: 39656908D
          - supply_point: "ES0022957057732644JV"
            supply_point_address: Cuesta Iglesia, 42
            supply_point_postalcode: 08066
            supply_point_city: Olvan
            supply_point_state: Barcelona
            distribution_coefficient: 0.5
            owner_name: Víctor
            owner_surname1: Puig
            owner_surname2: de la Serra
            owner_vat: 06454209H
        _links:
          self: https://somcomunitats.coop/api/energy_selfconsumption/projects/ES1234123456789012JY1FA000/members?page_size=2
          previous_page: null
          next_page: https://somcomunitats.coop/api/energy_selfconsumption/projects/ES1234123456789012JY1FA000/members?page=2&page_size=2
    badrequest-response-example:
      value:
        error:
          code: bad_request
          message: "Sorry, I couldn't process this request  ¯\_(ツ)_/¯"
    badauthrequest-response-example:
      value:
        error:
          code: bad_request
          message: Missing username or password.
    unauthorized-response-example:
      value:
        error:
          code: unauthorized
          message: Auth required.